requirements.txt
#Requirements

sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
langchain_community
langchain_openai
langchain_experimental
langchain_huggingface
-e .